requirements.txt
asgiref==3.8.1
Django==4.2.7
djangorestframework==3.15.1
python-dotenv==1.0.0
sqlparse==0.5.0
tzdata==2024.1
django-phonenumber-field==7.3.0
django_phonenumbers